test: Split up Zoo mocks to avoid vitest import error for expect.any()
