+++ editor.planx.uk/src/@planx/components/List/Public/tests/index.payloadGeneration.test.tsx
@@ -8,7 +8,8 @@ import {
   mockUnitsPayload,
   mockUnitsProps,
 } from "../../schemas/mocks/GenericUnits";
-import { mockZooPayload, mockZooProps } from "../../schemas/mocks/Zoo";
+import { mockZooPayload } from "../../schemas/mocks/Zoo/payload";
+import { mockZooProps } from "../../schemas/mocks/Zoo/props";
 import ListComponent from "..";
 import { fillInResponse } from "./testUtils";
 
@@ -48,23 +49,39 @@ describe("Payload generation", () => {
       await user.click(screen.getByTestId("continue-button"));
 
       expect(handleSubmit).toHaveBeenCalled();
-      const result: typeof mockZooPayload = handleSubmit.mock.calls[0][0]
+      const result: typeof mockZooPayload = handleSubmit.mock.calls[0][0];
       expect(result).toMatchObject(mockZooPayload);
 
       // All FileUploadField values are merged (2 files for each of the 2 list items)
       expect(result.data["photographs.existing"]).toHaveLength(4);
 
       // Schema "FileTypes" are extracted to the root level of the passport
-      expect(result.data["photographs.existing"][0].id).toEqual(result.data.mockFn[0]["photographs.existing"][0].id)
-      expect(result.data["photographs.existing"][1].id).toEqual(result.data.mockFn[0]["photographs.existing"][1].id)
-      expect(result.data["photographs.existing"][2].id).toEqual(result.data.mockFn[1]["photographs.existing"][0].id)
-      expect(result.data["photographs.existing"][3].id).toEqual(result.data.mockFn[1]["photographs.existing"][1].id)
+      expect(result.data["photographs.existing"][0].id).toEqual(
+        result.data.mockFn[0]["photographs.existing"][0].id,
+      );
+      expect(result.data["photographs.existing"][1].id).toEqual(
+        result.data.mockFn[0]["photographs.existing"][1].id,
+      );
+      expect(result.data["photographs.existing"][2].id).toEqual(
+        result.data.mockFn[1]["photographs.existing"][0].id,
+      );
+      expect(result.data["photographs.existing"][3].id).toEqual(
+        result.data.mockFn[1]["photographs.existing"][1].id,
+      );
 
       // Schema "FileTypes" are still kept as flattened values within the passport
-      expect(result.data["photographs.existing"][0].id).toEqual(result.data["mockFn.one.photographs.existing"][0].id)
-      expect(result.data["photographs.existing"][1].id).toEqual(result.data["mockFn.one.photographs.existing"][1].id)
-      expect(result.data["photographs.existing"][2].id).toEqual(result.data["mockFn.two.photographs.existing"][0].id)
-      expect(result.data["photographs.existing"][3].id).toEqual(result.data["mockFn.two.photographs.existing"][1].id)
+      expect(result.data["photographs.existing"][0].id).toEqual(
+        result.data["mockFn.one.photographs.existing"][0].id,
+      );
+      expect(result.data["photographs.existing"][1].id).toEqual(
+        result.data["mockFn.one.photographs.existing"][1].id,
+      );
+      expect(result.data["photographs.existing"][2].id).toEqual(
+        result.data["mockFn.two.photographs.existing"][0].id,
+      );
+      expect(result.data["photographs.existing"][3].id).toEqual(
+        result.data["mockFn.two.photographs.existing"][1].id,
+      );
     },
   );
