+++ editor.planx.uk/src/@planx/components/List/schemas/mocks/Zoo/props.ts
@@ -1,0 +1,10 @@
+import { Props } from "../../../Public";
+import { Zoo } from "./schema";
+
+export const mockZooProps: Props = {
+  fn: "mockFn",
+  schema: Zoo,
+  schemaName: "Zoo",
+  title: "Mock Title",
+  description: "Mock description",
+};
